requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
